+++ src/internal/isCharAtoZUppercase.ts
@@ -1,0 +1,4 @@
+/** @internal */
+export function isCharAtoZUppercase(value: number): boolean {
+  return value >= 65 /* A */ && value <= 90 /* Z */;
+}