+++ src/internal/isCharDigit.ts
@@ -1,0 +1,4 @@
+/** @internal */
+export function isCharDigit(value: number): boolean {
+  return value >= 48 /* 0 */ && value <= 57 /* 9 */
+}