+++ src/internal/removeWhitespaces.ts
@@ -1,0 +1,4 @@
+/** @internal */
+export function removeWhitespaces(value: string): string {
+  return value.replace(/\s/g, "");
+}