+++ src/types/ibanParseOptions.ts
@@ -1,0 +1,4 @@
+export interface IbanParseOptions {
+  removeWhitespaces: boolean;
+  validateStructure: boolean;
+}